--- src/pages/setting/index.tsx
@@ -1,14 +1,0 @@
-import { Suspense } from "react";
-import { Layout } from "src/components/layout";
-import { AppLoading } from "src/components/ui-libraries/AppLoading";
-
-const Setting = () => {
-  return (
-    <Layout>
-      <Suspense fallback={<AppLoading />}>
-        <h1>Setting</h1>
-      </Suspense>
-    </Layout>
-  );
-};
-export default Setting;
